Add files via upload
Create an abstract class, called Instrument which will contains a field name and a method called play,
that must be implemented by a sub-class.
You need to define a sub-class called StringedInstrument that extends the Instrument class and adds an
extra field called numberOfStrings. Then you add two more classes "ElectricGuitar and
ElectricBassGuitar" that implements the functionality of a StringedIntrument.